FOODSHARIN-96 refactor maps
