+++ services/frontend/src/components/maps/Locate.jsx
@@ -1,0 +1,24 @@
+import React from "react";
+
+const Locate = ({ panTo }) => {
+  return (
+    <Button
+      onClick={() => {
+        navigator.geolocation.getCurrentPosition(
+          (position) => {
+            panTo({
+              lat: position.coords.latitude,
+              lng: position.coords.longitude,
+            });
+          },
+          (error) => console.log(error),
+          options
+        );
+      }}
+    >
+      Current Location
+    </Button>
+  );
+};
+
+export default Locate;
